+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Next
+
+### Changed
+
+- `Raxx.set_header/2` raises an `ArgumentError` when setting host headers.
+- `ArgumentError` is raised instead of `RuntimeError` in cases of bad headers and body content.
+- `Raxx.set_body/2` raises an `ArgumentError` for GET and HEAD requests.
+
 ## [0.16.1](https://github.com/CrowdHailer/raxx/tree/0.16.1) - 2018-09-19
 
-## Fixed
+### Fixed
 
 - `Raxx.NotFound` was incorrectly building body for `handle_request` callback.